requirements.txt
torch
torchvision
opencv-python
streamlit
Pillow
tqdm
